+++ src/ISettlerActions.sol
@@ -6,7 +6,6 @@ import {ISignatureTransfer} from "@permit2/interfaces/ISignatureTransfer.sol";
 interface ISettlerActions {
     /// VIP actions should always start with `recipient` address and the `permit` from the taker
     /// followed by all the other parameters to ensure compatibility with `executeWithPermit` entrypoint.
-
     /// @dev Transfer funds from msg.sender Permit2.
     function TRANSFER_FROM(address recipient, ISignatureTransfer.PermitTransferFrom memory permit, bytes memory sig)
         external;
@@ -296,7 +295,8 @@ interface ISettlerActions {
         uint256 amountOutMin
     ) external;
 
-    function RENEGADE(address target, address sellToken, bool isSellingBase, bytes memory data) external;
+    function RENEGADE(address target, address sellToken, bool baseForQuote, uint256 minBuyAmount, bytes memory data)
+        external;
 
     struct BebopMakerSignature {
         bytes signatureBytes;