+++ test/integration/arbitrum/Renegade.t.sol
@@ -4,6 +4,10 @@ pragma solidity ^0.8.25;
 import {IERC20} from "@forge-std/interfaces/IERC20.sol";
 import {ARBITRUM_SELECTOR} from "src/core/Renegade.sol";
 import {ArbitrumSettler} from "src/chains/Arbitrum/TakerSubmitted.sol";
+import {ActionDataBuilder} from "../../utils/ActionDataBuilder.sol";
+import {ISettlerActions} from "src/ISettlerActions.sol";
+import {ISettlerBase} from "src/interfaces/ISettlerBase.sol";
+import {SettlerBasePairTest} from "../SettlerBasePairTest.t.sol";
 import {
     ARBITRUM_GAS_SPONSOR,
     ARBITRUM_TXN_CALLDATA,
@@ -12,15 +16,13 @@ import {
     ABRITRUM_WETH,
     ARBITRUM_AMOUNT
 } from "../RenegadeTxn.t.sol";
-import {RenegadeTest} from "../Renegade.t.sol";
 
-contract RenegadeArbitrumIntegrationTest is RenegadeTest {
+// Sell-quote: USDC -> WETH
+// Replays https://arbiscan.io/tx/0x937b0111b64ce852f1202c324753f6e1066d1137e5aee5d2a076997b3c873dec
+contract RenegadeArbitrumIntegrationTest is SettlerBasePairTest {
     function setUp() public virtual override {
-        target = ARBITRUM_GAS_SPONSOR;
-        txnCalldata = ARBITRUM_TXN_CALLDATA;
-        selector = ARBITRUM_SELECTOR;
-
         super.setUp();
+        vm.label(ARBITRUM_GAS_SPONSOR, "GasSponsor");
     }
 
     function settlerInitCode() internal virtual override returns (bytes memory) {
@@ -44,10 +46,50 @@ contract RenegadeArbitrumIntegrationTest is RenegadeTest {
     }
 
     function _testName() internal pure virtual override returns (string memory) {
-        return "USDC-WETH";
+        return "ARBITRUM-RENEGADE";
     }
 
     function amount() internal pure virtual override returns (uint256) {
         return ARBITRUM_AMOUNT;
+    }
+
+    function testSellQuote() public {
+        bytes memory _calldata = ARBITRUM_TXN_CALLDATA;
+        assembly ("memory-safe") {
+            let len := mload(_calldata)
+            _calldata := add(0x04, _calldata)
+            mstore(_calldata, sub(len, 4))
+        }
+
+        deal(address(fromToken()), address(this), amount());
+        fromToken().approve(address(allowanceHolder), amount());
+        allowanceHolder.exec(
+            address(settler),
+            address(fromToken()),
+            amount(),
+            payable(address(settler)),
+            abi.encodeCall(
+                settler.execute,
+                (
+                    ISettlerBase.AllowedSlippage({
+                        recipient: payable(address(this)), buyToken: toToken(), minAmountOut: 0
+                    }),
+                    ActionDataBuilder.build(
+                        abi.encodeCall(
+                            ISettlerActions.TRANSFER_FROM,
+                            (
+                                address(settler),
+                                defaultERC20PermitTransfer(address(fromToken()), amount(), 0),
+                                new bytes(0)
+                            )
+                        ),
+                        abi.encodeCall(
+                            ISettlerActions.RENEGADE, (ARBITRUM_GAS_SPONSOR, address(fromToken()), false, 0, _calldata)
+                        )
+                    ),
+                    bytes32(0)
+                )
+            )
+        );
     }
 }